WIP: Correct error distance computation.
When any motion of the structure is invalid, the program has to detect
this (easy) and compute the allowed distance, that if instead of perform
the required motion, we had performed this new distance, there will be
no error, and consequently, the structure would remain at the limit. It
is already done for advance, elevate and shift. Still not done for
incline.