Modifiy distance returned when the wheel is inside
This modification is useful when the wheel get inside a step a distance
greater that its radius. In this case, to take the wheel back to a valid
position moving it in horizontal direction, we need the distance added
to the function. This is not a very useful value, since this case is
impossible to happen if we move the structure by small distances. But it
was implemented just for checking reasons.