Remove parameter distance in all checking functions.
The parameter distance was initially added to these functions so that
the functions know the direnction of the motion, and so when more than
one error distance is detected, know if they must return the minimum or
the maximum of these errors. But it can easily see that always all
errors take the same sign, all positive or all negative, so that
comparing the absolute value or all errors are enough.