Organizing error distances in the structure.
Adding the distance_errors.py file, all the computations to get the
actual error to be returned when moving the structure and an error
(colllision or unstability) happens are performed in the classes
implemented in this file. So, the structure moving functions are now
more simple to understand.
Still to do:
 - Clean and comment all the code.
 - Run the unit_test to check that all the functions return the correct
error.