Adding functions for optimization modules.
Add the compute_time module. This module computes the total time
requirede to overcome a stair, without simulation, saving a big amount
of computation time. The module gets the dimensions of the structure and
stairs, and returns the total time (total number of iterations)
required. This is intended for the optimization function to execute one
structure in litte time.