Add horizontal margin for downstair direction
In this version, we I added a small margin when downstairs, so the
unstable wheel reaches the ground a little before the stable wheel get
out of the step. At the same time, I cleaned function get_distances from
wheel and pair modules, so that the margins are added in wheel module,
and not in pair module, which is more convenient. The figures in
get_distances folder has been also modified to reflect these changes.