New algorithm for control function.
Inside the compute_instruction function, extract the code for each wheel
to a new lis of functions (make_room_wheelX), where all the computations
needed when the basic motion is not enough to complete the actuator
motion, is performed. Now it is more clear to understand the algorithm.
Still to do, debug the last_instruction funcion, which is now working
bad.