Bug on control module
When the space between wheels of the same pair is large enogh so that
there is little space, the program raises a run time error. In this
case, the code has been corrected so that it can be detected in the
control module, meaning that this stair can not be climbed.