Change main loop behaviour
The following changes has been included:
- When the simulation finishes or crashes, the program change to manual
mode, so that the user can still move the structure, or finish the
program. In the previous version, the program finishes, which is
sometimes an incovenience, because we can not observe the cause of the
failiure.
- Delete the variable manual mode, and substitute it for the member of
the graphics, with already had the same value.