Allow for fractional samples when simulating.
In the previous version, when simulating we need to round the time
required to complete an instruction to the next integer. Since all this
fractions of samples get accumulated, the total time required in
simulation is always equal or normally greater than the actual time
required. In this version, it is allowed (with a simple trick) to
simulate with non integer sample times. See simulator.simulate_step ->
instructions after the call to self.compute_time(instruction) for more
info.