Simplify wheel.get_distances.
When detecting if the structure is facing and up or down stair, it is
not necessary to compare the previous step, only need to compare the
next step. This is a better approach if we implement the funcionality to
cross two stairs with different directions, although this is more
complicated (need to modificate pair.py), but the correction here is a
simplification of the function, which is good to commit.