Bug in simulator.compute_time
Update the current speed of the structure at the begining of the
function. Previously, this update were done within the simulate_step
function, but in some ocasions, the function is not called (when the
instruction time is lower than the sample time), and a exception were
raised in these cases. Now, a redundant update ensures that this error
never happens.