Version 2
Final code for version 2. The code includes the dynamics of the
horizontal motion, that is, it allows setting a finite maximum
acceleration and decceleration, and a maximum speed. The main work has
been done in profile.py file (see also documentation).

Also added the funcionality that, at the end of the stair, the structure
stop at the end of the landing space, not at the begining.