Add equal speed function to profiles.
Instead of comparing the values of two speeds to check both are similar
(are within a rounding error margin), make a function that do this
comparison. Apart from reducing code, it is important that wherever we
check a speed, this comparison always return the same, that is it uses
the same threshold in any point of code.