Change main loop
The loop is now more easy to understand.

TODO: When pressing tab key to exit manual mode, the system computes a new automatic instruction, and simulate the first step. That is, if we start pressing tab key continuously to enter and exit manual - automatic mode, the whole simulation wuold be done. In would be better to not do anything when exiting manual mode.