Optimize time computation.
If we use dynamic control, the module need to have some future
instructions computed in advance, to check for any crash if the end
speed of the current instruction is high enough. For that reason, the
same instruction could be computed many times. With this change in the
compute_distance function, we allow the system to prevent it from
computing the same instruction many times.