Change project structure.
Move all the simulator and control modules to the module "simulator".
Move all graphics modules to the module "graphics". This way, when we do
not need graphics functions, we do not have to load this module.

NOTE: This is the correct version. Tried some hours before, but the program crashes because I do not set correctly the interaction between Agg from matplotlib y waitKey from opencv.