Improve manual control of the structure.
Since the maual control only can make discrete step, it is possible
that, if the distance to move is less than the step, the motion can not
be done, because we are asking the structure more distance than the
remaining one. For example, if we need one wheel to contact the step,
but the distance between the wheel and the step is less than the
discrete motion for the actuator shift, the wheel can never be place in
the step.

In this case, we only modify the actuator shift and the structure
advance, since the other motions are not needed so far. In case we need
to update the rest of the motions, just copy the added code.