Change csv generation
The csv is used to print actuator position along the time. This is
needed to be compatible con Inventor. In this commit, the format is one
csv file for actuator, with two columns, one for the time, and the other
for the actuator position. We need six files, one for each actuator, and
anothe for the structure speed.