Change error distance objects
In the previous version, the collision and unstability errors were not
stores appropriately, and it was difficult to infer the error from the
information we get. In this new version, ALL the errors are stored in a
new class (StructureError class), and this class has also all the method
needed to infer which error was, and which correction is needed (this is
specially important for structure inclination, were a lot of different
kind of errors can happen).