Check collisions when inclining fixing any of the actuators
In the previous version, the structure could only incline fixing the
rear of the front actuator. In the new version, the structure can
incline fixing any of the four actuator. Here, all the code to detect
collisions have been modified so that the collision can be computed
correctly independently the actuator we are fixing.