Update push actuator
Push actuator is a new function that, when trying to shift an actuator,
and there is no room for the actuator to complete the motion, this
function, along with many other functions, try to make more room by
inclining - elevating the structure, following a given algorithm. This
algorithm is:
- For actuator 3, first incline from the front, and if not possible, try
to incline from the rear.
- For the rest of the actuators, first try to elevate, and if not
possible, try to incline fixing the algorithm that has collided when
elevating.