Problem when pushing the front actuator collliding with act 1 and 2.
When inclining the structure when pushing the front actuator, if the
structure collides with actuator 1 and 2, it is possible that the module
can not make enough space for the front actuator, even though there is
enough space. See test_make_room_wheel (test motion10t and motion11t).
The problem is still not solved.