Remove ControlError exception.
In the previous version, we use the ControlError exception to tell that
the control module can not find a valid instruction. In this version,
the module does not raise any exception, just send the instruction that
the module has found, and probably in the next iteration, if the motion
is impossible, the module will generate a null instruction, which is
enough for the calling function to know that it is not possible to cross
the stair.

This is done this way, since, for some occasions where the module can
not find a valid instruction to perform the whole motion, it is possible
that in the next itaration, continuing from where the structure was
left, it can find a new instruction to continue. This is an improvement
from the previous version.