Update compute_time to the new control algorithm
In the new algorithm, the control module does not raise any exception
when it can not find a valid instruction. It just send the instruction
that actually can do, and probably, in the next iteration a new
instruction will be generated, so that the calling function knows that
the structure can not cross the stair.

To sign when the end of the stair has been reached, the control module
includes the key "end:True" to the instruction dictionary.

Both modification has been taken into account in the compute_time
function.